implement parameter checks in policies
The methods create_with? and update_with? can now be used to perform
validations on the parameters.
Currently the Domain policy checks whether a request contains an id or a
user_id and validates them. Thus a user must not assign a domain to
another user and a reseller can only assign domains to its own customers
and nobody else.

This is obviously prevents a lot of nasty situations that might have
happened before.

Ref T20

Signed-off-by: Thore Boedecker <me@foxxx0.de>